Refactored into better class names
